Move dot function from plasma.py to simulation.py
